requirements.txt
langchain==0.0.325
openai==0.28.1
python-dotenv==1.0.0
streamlit==1.28.1
streamlit-chat==0.1.1
youtube-transcript-api==0.6.1
pytube==15.0.0
urllib3==1.26.18